requirements.txt
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
azure-storage-blob==12.9.0